Update Remove and Edit Modals
