--- frontend/components/admin/components/RemoveReviewModal.tsx
@@ -1,11 +1,0 @@
-const RemoveReviewModal = () => {
-	return (
-		<div className="ml-4" data-testid="remove-review-modal-1">
-			<h2>
-				Are you sure you want to remove this review? This cannot be undone.
-			</h2>
-		</div>
-	)
-}
-
-export default RemoveReviewModal
