+++ frontend/components/admin/sections/FlaggedReviews.tsx
@@ -1,33 +1,21 @@
 import Alert from '@/components/alerts/Alert'
-import Modal from '@/components/modal/Modal'
 import {Review} from '@/util/interfaces/interfaces'
 import {useEffect, useState} from 'react'
 import useSWR, {useSWRConfig} from 'swr'
-import EditReviewModal from '../components/EditReviewModal'
-import RemoveReviewModal from '../components/RemoveReviewModal'
 import {fetcher} from '@/util/helpers/fetcher'
+import EditReviewModal from '@/components/modal/EditReviewModal'
+import RemoveReviewModal from '@/components/modal/RemoveReviewModal'
 
 const FlaggedReviews = () => {
 	const {mutate} = useSWRConfig()
 	const [editReviewOpen, setEditReviewOpen] = useState(false)
 	const [selectedReview, setSelectedReview] = useState<Review | undefined>()
-	const [newReview, setNewReview] = useState('')
 
 	const [flaggedReviews, setFlaggedReviews] = useState<Array<Review>>([])
 
 	const [removeReviewOpen, setRemoveReviewOpen] = useState(false)
 	const [success, setSuccess] = useState(false)
 	const [removeAlertOpen, setRemoveAlertOpen] = useState(false)
-
-	const [landlord, setLandlord] = useState<string>(
-		selectedReview?.landlord || '',
-	)
-	const [country, setCountry] = useState<string>(
-		selectedReview?.country_code || '',
-	)
-	const [city, setCity] = useState<string>(selectedReview?.city || '')
-	const [province, setProvince] = useState<string>(selectedReview?.state || '')
-	const [postal, setPostal] = useState<string>(selectedReview?.zip || '')
 
 	// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
 	const {data: reviews, error} = useSWR<Array<Review>>(
@@ -43,84 +31,8 @@ const FlaggedReviews = () => {
 		}
 	}, [reviews])
 
-	useEffect(() => {
-		if (selectedReview) {
-			setLandlord(selectedReview.landlord)
-			setCountry(selectedReview.country_code)
-			setCity(selectedReview.city)
-			setProvince(selectedReview.state)
-			setPostal(selectedReview.zip)
-			setNewReview(selectedReview.review)
-		}
-	}, [selectedReview])
-
 	if (error) return <div>failed to load</div>
 	if (!reviews) return <div>loading...</div>
-
-	const onSubmitRemoveReview = (id: number) => {
-		fetch('/api/delete-review', {
-			method: 'POST',
-			headers: {
-				'Content-Type': 'application/json',
-			},
-			body: JSON.stringify({id: id}),
-		})
-			.then((result) => {
-				if (!result.ok) {
-					throw new Error()
-				}
-			})
-			.then(() => {
-				mutate('/api/get-flagged').catch((err) => console.log(err))
-				setRemoveReviewOpen(false)
-				setSuccess(true)
-				setRemoveAlertOpen(true)
-			})
-			.catch((err) => {
-				console.log(err)
-				setRemoveAlertOpen(false)
-				setSuccess(false)
-				setRemoveAlertOpen(true)
-			})
-	}
-
-	const onSubmitEditReview = (id?: number) => {
-		const editedReview = {
-			...selectedReview,
-			landlord: landlord,
-			country: country,
-			city: city,
-			state: province,
-			zip: postal,
-			review: newReview,
-			admin_edited: true,
-			admin_approved: true,
-			flagged: false,
-		}
-		fetch('/api/edit-review', {
-			method: 'POST',
-			headers: {
-				'Content-Type': 'application/json',
-			},
-			body: JSON.stringify(editedReview),
-		})
-			.then((result) => {
-				if (!result.ok) {
-					throw new Error()
-				}
-			})
-			.then(() => {
-				mutate('/api/get-flagged').catch((err) => console.log(err))
-				setEditReviewOpen(false)
-				setSuccess(true)
-				setRemoveAlertOpen(true)
-			})
-			.catch((err) => {
-				console.log(err)
-				setSuccess(false)
-				setRemoveAlertOpen(true)
-			})
-	}
 
 	const onSubmitApproveReview = (review: Review) => {
 		const editedReview = {
@@ -161,40 +73,21 @@ const FlaggedReviews = () => {
 			) : null}
 			{selectedReview ? (
 				<>
-					<Modal
-						title="Edit Review"
-						open={editReviewOpen}
-						setOpen={setEditReviewOpen}
-						element={
-							<EditReviewModal
-								selectedReview={selectedReview}
-								landlord={landlord}
-								setLandlord={setLandlord}
-								country={country}
-								setCountry={setCountry}
-								city={city}
-								setCity={setCity}
-								province={province}
-								setProvince={setProvince}
-								postal={postal}
-								setPostal={setPostal}
-								review={newReview}
-								setReview={setNewReview}
-							/>
-						}
-						onSubmit={onSubmitEditReview}
-						selectedId={selectedReview?.id}
-						buttonColour="blue"
+					<EditReviewModal
+						selectedReview={selectedReview}
+						mutateString="/api/get-flagged"
+						setEditReviewOpen={setEditReviewOpen}
+						setSuccess={setSuccess}
+						setRemoveAlertOpen={setRemoveAlertOpen}
+						editReviewOpen={editReviewOpen}
 					/>
-					<Modal
-						title="Remove Review"
-						open={removeReviewOpen}
-						setOpen={setRemoveReviewOpen}
-						element={<RemoveReviewModal />}
-						onSubmit={onSubmitRemoveReview}
-						buttonColour="red"
-						selectedId={selectedReview?.id}
-					/>
+					<RemoveReviewModal selectedReview={selectedReview}
+						mutateString={'/api/get-flagged'}
+						setRemoveReviewOpen={setRemoveReviewOpen}
+						setSuccess={setSuccess}
+						setRemoveAlertOpen={setRemoveAlertOpen}
+						removeReviewOpen={removeReviewOpen} />
+					
 				</>
 			) : null}
 			<div className="container -mx-4 overflow-hidden shadow ring-1 ring-black ring-opacity-5 sm:-mx-6 md:mx-0 md:rounded-lg">
